requirements.txt
Django>=5.0,<6.0
djangorestframework
channels
channels-redis
celery
amqp
psycopg2-binary
django-cors-headers
flower
python-decouple
gdal==3.6.2
daphne
djangorestframework-simplejwt
django-celery-results
osmnx 
networkx 
geopandas
scikit-learn
Pillow
firebase-admin